main-app/requirements.txt
streamlit==1.52.2
python-docx==1.2.0
PyMuPDF==1.26.7